+++ database/migrations/2025_02_13_082227_create_hierarchy_level_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->integer('updated_by')->nullable();
             $table->integer('deleted_date')->nullable();
             $table->integer('deleted_by')->nullable();
-            $table->integer('deleted')->default(0);
+            $table->unsignedTinyInteger('deleted')->default(0);
         });
     }
